+++ main1067.cpp
@@ -35,7 +35,7 @@ int main(void) {
 	vEYrpmo	=	(short)-346684528;
 	yD8KDn	=	(short)-380043828;
 	x1unm	=	(short)-1453452398;
-	CVN3	=	-1401595523;
+	CVN3	=	(short)-1401595523;
 	tC2Gyjc2Og	=	972863904;
 	AcuH3O2K	=	1745005715;
 	XBA_6WS4	=	1203282286;
@@ -45,7 +45,7 @@ int main(void) {
 	o4haXKFCT	=	(short)-2112031713;
 	WambQ08	=	(short)-1712610904;
 	m_0rv	=	(short)366653337;
-	hp43	=	1501292720;
+	hp43	=	(short)1501292720;
 	R87A7b	=	-1624702077;
 	i1cD	=	-281874110;
 	yx7_	=	-418241679;
@@ -55,7 +55,7 @@ int main(void) {
 	kW9s5KZ6	=	(short)536496236;
 	M8716CBy	=	(short)698778325;
 	lw3mQS	=	(short)-368646700;
-	E1ex6e	=	785795512;
+	E1ex6e	=	(short)785795512;
 	nvRe7vSzj	=	247380341;
 	eYuennSs04	=	2110630737;
 	GU45	=	1026047542;
@@ -65,7 +65,7 @@ int main(void) {
 	xN4p855	=	(short)-1009016081;
 	gRB_xEpY6e	=	(short)-1025684480;
 	a82g5H7613	=	(short)1196837668;
-	vtoL2	=	911720013;
+	vtoL2	=	(short)911720013;
 	E48Eo43f	=	785251614;
 	zoPcj	=	-1523691490;
 	q_1_	=	457561723;
@@ -75,7 +75,7 @@ int main(void) {
 	yg0N1hh4	=	(short)1787705397;
 	noBdxx	=	(short)-483017923;
 	d2A855X	=	(short)-213125608;
-	Bxe9dq	=	76373843;
+	Bxe9dq	=	(short)76373843;
 
 nvRe7vSzj = itW7nO9_v3(nvRe7vSzj);
 Bxe9dq = o4haXKFCT / Qt9_9;
