+++ main1085.cpp
@@ -33,7 +33,7 @@ int main() {
  dpV1L = (short)-1044462076;
  u0X9zG_ = (short)387550236;
  Gv42c1 = (short)1816881915;
- nnPhr = -331422951;
+ nnPhr = (short)-331422951;
  yIcfrD = 1889396305;
  I050 = -1430502584;
  bTIKpmS1Q = -624178730;
@@ -43,7 +43,7 @@ int main() {
  sS4k7gI = (short)1541451526;
  L86cOV2YP = (short)2008883426;
  g4dvk00 = (short)1415274989;
- AFfq1F = 1930457664;
+ AFfq1F = (short)1930457664;
  a8RLU = -870031138;
  G2An = -560467518;
  E6_A9179 = -501626062;
@@ -53,7 +53,7 @@ int main() {
  ZL7q4ZEa6 = (short)-70063970;
  klkgYh9sBt = (short)-1521629974;
  es61KiVQ1N = (short)-1327904753;
- b24R3CK = 350247981;
+ b24R3CK = (short)350247981;
  pO02eu9a44 = 1421060024;
  W513_wc = -987579613;
  VSsWyB5A9 = -814197120;
@@ -63,7 +63,7 @@ int main() {
  s3HVu66Aq = (short)1816316830;
  R9O5j = (short)-701128282;
  P5339 = (short)1430671142;
- P4t68L6IXN = -737725794;
+ P4t68L6IXN = (short)-737725794;
  ycWkXH = 565518433;
  FeUPIBG = 450803022;
  L_B_Y517T = 858634871;
@@ -73,7 +73,7 @@ int main() {
  COjywZ_F7F = (short)-965716756;
  g390bqdy = (short)-1847213074;
  i6CwX7w9 = (short)1522621973;
- W23S9 = 1393420448;
+ W23S9 = (short)1393420448;
  vGarCg2 = 1047663072;
  WseZjIi = 452715268;
  l2_060pC09 = -913658366;
@@ -83,7 +83,7 @@ int main() {
  i714_S6X_6 = (short)877018803;
  qn39gTJ0R = (short)-387107563;
  QkIM = (short)-468161275;
- I9heDJ1 = -1970601490;
+ I9heDJ1 = (short)-1970601490;
  toy1Pj_K = -1573921277;
  k644Pcm8 = -354866793;
  J_yW3 = 2129776826;
@@ -93,7 +93,7 @@ int main() {
  Hqe9RcV2 = (short)426439464;
  a2K21bo2C0 = (short)71828675;
  s7xPS9 = (short)-139434621;
- bSjy = -125975361;
+ bSjy = (short)-125975361;
  ydL5 = -484805139;
  s4RovG = -190975023;
  jjX_U4 = 749384205;
@@ -103,7 +103,7 @@ int main() {
  yhfFSI4_ = (short)-2117678778;
  h_i81 = (short)1320261483;
  W7X6dW = (short)1711554220;
- k61A_39tl = 1717900049;
+ k61A_39tl = (short)1717900049;
  U5Pt8 = 903810294;
  lZ_PMH8G = 432242111;
  C3tj9tZh_ = 1571944737;
@@ -113,7 +113,7 @@ int main() {
  q34Yx8 = (short)-548063351;
  B_8_qD4wZ = (short)903306795;
  s8AiXD = (short)378731905;
- L832J9g = 1614951635;
+ L832J9g = (short)1614951635;
  uR_Y = 1739906990;
  Sn4H17AFJq = 768758372;
  u4H494 = -985225885;
