+++ main1094.cpp
@@ -16,7 +16,7 @@ int main(int argc, char** argv) {
     q_5Lt8qGo    =    (unsigned short)-1631699966;
     Ql3zoZ_r    =    (unsigned short)1819448592;
     du0a6nlQz0    =    (unsigned short)-1799762045;
-    CpXi    =    1827354453;
+    CpXi    =    (unsigned short)1827354453;
     K8xB_J52pA    =    987088401;
     J9wY    =    1678101638;
     w9U7jov81    =    -1906441278;
